[pkg/stanza] Cleanup add operator files
