+++ pkg/stanza/operator/transformer/add/config.go
@@ -4,11 +4,9 @@
 package add // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/transformer/add"
 
 import (
-	"context"
 	"fmt"
 	"strings"
 
-	"github.com/expr-lang/expr/vm"
 	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
@@ -67,40 +65,4 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 
 	addOperator.program = compiled
 	return addOperator, nil
-}
-
-// Transformer is an operator that adds a string value or an expression value
-type Transformer struct {
-	helper.TransformerOperator
-
-	Field   entry.Field
-	Value   any
-	program *vm.Program
-}
-
-// Process will process an entry with a add transformation.
-func (p *Transformer) Process(ctx context.Context, entry *entry.Entry) error {
-	return p.ProcessWith(ctx, entry, p.Transform)
-}
-
-// Transform will apply the add operations to an entry
-func (p *Transformer) Transform(e *entry.Entry) error {
-	if p.Value != nil {
-		return e.Set(p.Field, p.Value)
-	}
-	if p.program != nil {
-		env := helper.GetExprEnv(e)
-		defer helper.PutExprEnv(env)
-
-		result, err := vm.Run(p.program, env)
-		if err != nil {
-			return fmt.Errorf("evaluate value_expr: %w", err)
-		}
-		return e.Set(p.Field, result)
-	}
-	return fmt.Errorf("add: missing required field 'value'")
-}
-
-func isExpr(str string) bool {
-	return strings.HasPrefix(str, "EXPR(") && strings.HasSuffix(str, ")")
 }
