+++ pkg/stanza/operator/transformer/add/transformer_test.go
